Tiling of sprite images
Sprite images when dragged to the canvas are now tiled instead of simply
stretched to whatever dimension the user sets. This ensures original
aspect ratio of image is maintained, and is simply repeated when the
sprite gets too wide. May be necessary in future to allow tiling in
either direction, and turn it on or off.